+++ src/policy/policy.cpp
@@ -241,12 +241,12 @@ int64_t GetTransactionSize(int64_t nSize, int64_t nSigOpCost)
     return std::max(nSize, nSigOpCost * nBytesPerSigOp);
 }
 
-int64_t GetTransactionSize(const CTransaction& tx, int64_t nSigOpCost)
+int64_t GetTransactionSize(const CTransaction& tx, int32_t nSigOpCost)
 {
     return GetTransactionSize(GetSerializeSize(tx, SER_NETWORK, PROTOCOL_VERSION), nSigOpCost);
 }
 
-int64_t GetTransactionInputSize(const CTxIn& txin, int64_t nSigOpCost)
+int64_t GetTransactionInputSize(const CTxIn& txin, int32_t nSigOpCost)
 {
     return GetTransactionSize(GetSerializeSize(txin, SER_NETWORK, PROTOCOL_VERSION), nSigOpCost);
 }
